+++ cookbooks/groupware/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file is used to list changes made in each version of groupware.
 
+## 0.4.0:
+
+* recipes/default - now includes mumble and xchat install management
+* recipes/mumble  - installs Mumble low latency VoIP client
+* recipes/xchat   - installs XChat IRC client
+
 ## 0.3.0:
 
 * recipe/default  - now checks for section presence in box profile's apps list before proceed